extract all network code to legacy submodule
